refactor(tests): split E2eReal into E2ePublic + E2eCredentials (3-tier E2E)
Tier 1 — E2eMocked: no bank access, always runs (unchanged)
Tier 2 — E2ePublic: real bank URLs, wrong creds, no secrets needed → always runs
  - SelectorFallback* tests (selector fallback with invalid creds)
  - InvalidLoginBatch* tests (wrong credentials detection for all banks)
Tier 3 — E2eCredentials: real bank secrets required → local (.env) + CI secrets
  - Amex, Discount, Isracard, Max, VisaCal, Beinleumi, OneZero

CI (e2e.yml):
  - e2e-public: always runs (no condition), catches form/selector regressions
  - e2e-credentials: internal PRs only (has secrets), catches login regressions

Pre-commit hook:
  - Tier 2 (E2ePublic) always runs after Mocked — no credentials required
  - Tier 3 (E2eCredentials) only when .env exists (existing behavior)

Scripts added: test:e2e:public, test:e2e:credentials (replaces test:e2e:real)
