+++ .husky/pre-commit
@@ -28,6 +28,11 @@ echo "🎭  [3/8] Mocked E2E tests..." | tee -a "$LOG_FILE"
 run_gate npm run test:e2e:mock --no-coverage || exit 1
 
 echo "" | tee -a "$LOG_FILE"
+echo "🏛️  [3b] Public E2E tests (selector fallback + invalid login — no credentials needed)..." | tee -a "$LOG_FILE"
+npx ncp src/Tests/.tests-config.tpl.js src/Tests/.tests-config.js 2>/dev/null
+run_gate npm run test:e2e:public -- --no-coverage || exit 1
+
+echo "" | tee -a "$LOG_FILE"
 echo "🧹  [4/8] Lint + format (CI parity)..." | tee -a "$LOG_FILE"
 run_gate npx lint-staged || exit 1
 echo "  → ESLint all src/ (matches CI: npx eslint src --max-warnings 0)..." | tee -a "$LOG_FILE"
@@ -47,14 +52,14 @@ echo "🔒  [6/8] npm audit (production deps)..." | tee -a "$LOG_FILE"
 run_gate npm audit --audit-level=high --omit=dev || exit 1
 
 echo "" | tee -a "$LOG_FILE"
-echo "🏦  [7/8] Real E2E — non-OTP banks (Amex, Isracard, Max, Discount, VisaCal)..." | tee -a "$LOG_FILE"
+echo "🏦  [7/8] Credential E2E — non-OTP banks with real credentials (Amex, Isracard, Max, Discount, VisaCal)..." | tee -a "$LOG_FILE"
 if [ -f .env ]; then
   npx ncp src/Tests/.tests-config.tpl.js src/Tests/.tests-config.js 2>/dev/null
   run_gate npx jest --no-coverage --testPathIgnorePatterns='/node_modules/' --testPathIgnorePatterns='E2eMocked' \
-    --testPathPatterns='E2eReal/Amex|E2eReal/Isracard|E2eReal/Max|E2eReal/Discount|E2eReal/Visacal' \
+    --testPathPatterns='E2eCredentials/Amex|E2eCredentials/Isracard|E2eCredentials/Max|E2eCredentials/Discount|E2eCredentials/Visacal' \
     --verbose || exit 1
 else
-  echo "  ⚠️  No .env file — skipping real E2E (CI will run them)" | tee -a "$LOG_FILE"
+  echo "  ⚠️  No .env file — skipping credential E2E (CI will run them with secrets)" | tee -a "$LOG_FILE"
 fi
 
 echo "" | tee -a "$LOG_FILE"
