+++ jest.config.js
@@ -8,19 +8,16 @@ const config = {
   rootDir: './src',
   transform: {
     '^.+\\.ts$': ['ts-jest'],
-    '@faker-js.+\\.js$': ['ts-jest', { diagnostics: false, tsconfig: { allowJs: true, checkJs: false } }],
+    '@faker-js.+\\.js$': [
+      'ts-jest',
+      { diagnostics: false, tsconfig: { allowJs: true, checkJs: false } },
+    ],
   },
   transformIgnorePatterns: ['/node_modules/(?!@faker-js/faker)'],
-  setupFilesAfterEnv: [
-    './Tests/JestSetup.ts',
-  ],
+  setupFilesAfterEnv: ['./Tests/JestSetup.ts'],
   testEnvironment: 'node',
-  testPathIgnorePatterns: ['/node_modules/', 'E2eReal/', 'E2eMocked/'],
-  collectCoverageFrom: [
-    '**/*.ts',
-    '!Tests/**',
-    '!**/*.test.ts',
-  ],
+  testPathIgnorePatterns: ['/node_modules/', 'E2ePublic/', 'E2eCredentials/', 'E2eMocked/'],
+  collectCoverageFrom: ['**/*.ts', '!Tests/**', '!**/*.test.ts'],
   coverageThreshold: {
     global: {
       branches: 76,
