+++ src/Tests/E2eCredentials/Beinleumi.e2e-credentials.test.ts
@@ -10,7 +10,7 @@ import {
   lastMonthStartDate,
   logScrapedTransactions,
   SCRAPE_TIMEOUT,
-} from './Helpers';
+} from '../E2ePublic/Helpers';
 
 dotenv.config();
