+++ src/Tests/E2eCredentials/Discount.e2e-credentials.test.ts
@@ -8,33 +8,33 @@ import {
   lastMonthStartDate,
   logScrapedTransactions,
   SCRAPE_TIMEOUT,
-} from './Helpers';
+} from '../E2ePublic/Helpers';
 
 dotenv.config();
 
 const hasCredentials = !!(
-  process.env.AMEX_ID &&
-  process.env.AMEX_CARD6DIGITS &&
-  process.env.AMEX_PASSWORD
+  process.env.DISCOUNT_ID &&
+  process.env.DISCOUNT_PASSWORD &&
+  process.env.DISCOUNT_NUM
 );
 const DESCRIBE_IF = hasCredentials ? describe : describe.skip;
 
-DESCRIBE_IF('E2E: Amex (real credentials)', () => {
+DESCRIBE_IF('E2E: Discount Bank (real credentials)', () => {
   beforeAll(() => {
     jest.setTimeout(SCRAPE_TIMEOUT);
   });
 
   it('scrapes transactions successfully', async () => {
     const scraper = createScraper({
-      companyId: CompanyTypes.Amex,
+      companyId: CompanyTypes.Discount,
       startDate: lastMonthStartDate(),
       shouldShowBrowser: false,
       args: BROWSER_ARGS,
     });
     const result = await scraper.scrape({
-      id: process.env.AMEX_ID ?? '',
-      card6Digits: process.env.AMEX_CARD6DIGITS ?? '',
-      password: process.env.AMEX_PASSWORD ?? '',
+      id: process.env.DISCOUNT_ID ?? '',
+      password: process.env.DISCOUNT_PASSWORD ?? '',
+      num: process.env.DISCOUNT_NUM ?? '',
     });
 
     assertSuccessfulScrape(result);
@@ -43,16 +43,12 @@ DESCRIBE_IF('E2E: Amex (real credentials)', () => {
 
   it('fails with invalid credentials', async () => {
     const scraper = createScraper({
-      companyId: CompanyTypes.Amex,
+      companyId: CompanyTypes.Discount,
       startDate: new Date(),
       shouldShowBrowser: false,
       args: BROWSER_ARGS,
     });
-    const result = await scraper.scrape({
-      id: '000000000',
-      card6Digits: '000000',
-      password: 'invalid123',
-    });
+    const result = await scraper.scrape({ id: '000000000', password: 'invalid123', num: '000000' });
     assertFailedLogin(result);
   });
 });
