+++ src/Tests/E2eCredentials/Max.e2e-credentials.test.ts
@@ -8,7 +8,7 @@ import {
   lastMonthStartDate,
   logScrapedTransactions,
   SCRAPE_TIMEOUT,
-} from './Helpers';
+} from '../E2ePublic/Helpers';
 
 dotenv.config();
