+++ src/Tests/E2eCredentials/OneZero.e2e-credentials.test.ts
@@ -12,7 +12,7 @@ import {
   lastMonthStartDate,
   logScrapedTransactions,
   SCRAPE_TIMEOUT,
-} from './Helpers';
+} from '../E2ePublic/Helpers';
 
 dotenv.config();
